Don't swallow query failure exceptions.
Printing exceptions and continuing during the build process is
pointless.  Unless someone is watching the (fairly noisy, which, to be
fair is on me) output, the error will go unnoticed.  Any automated build
process will complete, seemingly successfully.

Given the desired outcome is a coherently enriched KB, it seems
preferable to fail outright if the rules are not properly applied than
to give the impression that everything completed as expected.