Reorganized rules into pre- and post-identifier-merging groups
This is a first attempt at providing some inherent ordering to the
rules to indicate when a rule should be run relative to other rules and
relative to the identifier-merging step which is not rule-based. Along
with directories for the pre- and post-identifier-merging rules, a
deprecated folder has been populated with rules that are no longer
necessary.