Added bnode namespace as workaround for http:/ bug
The KR rule machinery appears to require the namespace of any URI it
processes to exist in the kabob.core.namespace file. If a rule
processes a URI that does not have a namespace reference in
kabob.core.namespace, then the URI will be corrupted, and will take the
form of http:/uri/goes/here, instead of http://uri/goes/here. Note the
missing slash (http:/)

Also added a test that is really a placeholder for testing the fix to
the http:/ bug.