Modified transfer_direct_links rule
Trying to debug a Exception in thread "main"
org.openrdf.repository.http.HTTPQueryEvaluationException: Bad chunk
size: HTTP/1.1 400  Bad Request.

The final line in the output file is for an ObjectProperty. It’s likely
that the two are unrelated, but I’ve excluded ObjectProperty’s from the
transfer_direct_link rule as they are already handled by a different
rule.