Migrated a script from kabob-base.docker to download irefweb and the reactive biopic file
Although most files consumed by KaBOB are downloaded automatically by
the datasource-fileparser code, the IRefWeb files have proven difficult
to retrieve through the file parser API. This script uses wget to
download the IRefWeb file.

There is no file parser for the Reactome biopax format as it is loaded
directly into the triple store, so this script also serves as a means
for downloading non-parsed files.