Changes made during development of the July 2018 release
* Fixed redundant OWL constructs when importing ontology blank nodes
* Added links from UniProt isoforms to their ‘canonical’ protein using variant_of
* Added labels to every (I think) exhaustive subclass that is created by the kabob rules

Note 1: There are still some nodes missing labels. Some are discontinued (NCBI) or withdrawn (HGNC) records that are linked by other sources. But many are genes/proteins from species other than human that are being brought in as part of PPIs, e.g. a human protein and a mouse protein are known to interact. The next build will restrict PPIs to just human-human interactions (I thought this was already the case but evidently it was not).

Note 2: There are still some redundant restriction classes. When there is a restriction that is defined in an ontology that is also defined by one of the KaBOB rules, e.g. only_in_taxon restrictions, there will be two copies b/c of the way the URIs are currently generated. I’ll work towards collapsing these in future releases.